requirements.txt
git+https://github.com/scottcarpenter28/Ambient-Weather-API-Connector.git
SQLAlchemy
python-dotenv